requirements.txt
# Core
requests
PyYAML
Flask

# AI & Machine Learning
tensorflow
scikit-learn

# Text-to-Speech
gTTS
elevenlabs

# Video Processing
moviepy

# Database
SQLAlchemy

# Telegram
python-telegram-bot